requirements.txt
numpy
pandas
scikit-learn
matplotlib==3.5.3
MySQL-python